+++ agent_build_refactored/managed_packages/steps/build_python_package_root.sh
@@ -53,7 +53,7 @@ mkdir -p "${OPENSSL_LIBS_DIR}"
 
 
 # This function copies Python's ssl module related files.
-function copy_and_patch_openssl_libs() {
+function copy_openssl_libs() {
   local openssl_libs_dir="$1"
   local python_step_output_dir="$2"
   local dst_dir="$3"
@@ -62,15 +62,12 @@ function copy_and_patch_openssl_libs() {
   libssl_path="$(find "${openssl_libs_dir}" -name "libssl.so.*")"
   local libcrypto_path
   libcrypto_path="$(find "${openssl_libs_dir}" -name "libcrypto.so.*")"
-  local libssl_filename
-  libssl_filename="$(basename "${libssl_path}")"
-  local libcrypto_filename
-  libcrypto_filename="$(basename "${libcrypto_path}")"
 
-  mkdir -p "${dst_dir}"
+  local libs_dir="${dst_dir}/libs"
+  mkdir -p "${libs_dir}"
   # Copy shared objects and other files of the OpenSSL library.
-  cp "${libssl_path}" "${dst_dir}"
-  cp "${libcrypto_path}" "${dst_dir}"
+  cp "${libssl_path}" "${libs_dir}"
+  cp "${libcrypto_path}" "${libs_dir}"
 
   local python_step_bindings_dir="${python_step_output_dir}${INSTALL_PREFIX}/lib/python${PYTHON_SHORT_VERSION}/lib-dynload"
 
@@ -84,46 +81,39 @@ function copy_and_patch_openssl_libs() {
 
   # Copy original ssl and hashlib C bindings. They will be used in case if appropriate OpenSSL version is
   # found on system.
-  local original_bindings_dir="${bindings_dir}/original"
-  mkdir -p "${original_bindings_dir}"
-  cp "${ssl_binding_path}" "${original_bindings_dir}"
-  cp "${hashlib_binding_path}" "${original_bindings_dir}"
-
-  # In case if there is no appropriate system OpenSSL, we also copy the same C bindings, but which are hardcoded
-  # to use OpenSSL shared objects that are shipped with the package.
-  local patched_bindings_dir="${bindings_dir}/patched"
-  mkdir -p "${patched_bindings_dir}"
-  local patched_ssl_binding_path
-  patched_ssl_binding_path="${patched_bindings_dir}/$(basename "${ssl_binding_path}")"
-  local patched_hashlib_binding_path
-  patched_hashlib_binding_path="${patched_bindings_dir}/$(basename "${hashlib_binding_path}")"
-  cp "${ssl_binding_path}" "${patched_bindings_dir}"
-  cp "${hashlib_binding_path}" "${patched_bindings_dir}"
-
-  new_dependencies_dir="$(realpath --relative-to "${PACKAGE_ROOT}" "${dst_dir}")"
-  # Patch ssl and hashlib C bindings and hardcode package's shared objects as dependencies.
-  patchelf --replace-needed "${libssl_filename}" "/${new_dependencies_dir}/${libssl_filename}" "${patched_ssl_binding_path}"
-  patchelf --replace-needed "${libcrypto_filename}" "/${new_dependencies_dir}/${libcrypto_filename}" "${patched_ssl_binding_path}"
-  patchelf --replace-needed "${libcrypto_filename}" "/${new_dependencies_dir}/${libcrypto_filename}" "${patched_hashlib_binding_path}"
+  mkdir -p "${bindings_dir}"
+  cp "${ssl_binding_path}" "${bindings_dir}"
+  cp "${hashlib_binding_path}" "${bindings_dir}"
 }
 
 # Copy ssl modules and libraries which are compiled for OpenSSL 1.1.1
-copy_and_patch_openssl_libs "${BUILD_OPENSSL_1_1_1}${LIBSSL_DIR}" "${BUILD_PYTHON_WITH_OPENSSL_1_1_1}" "${OPENSSL_LIBS_DIR}/1_1_1"
+copy_openssl_libs "${BUILD_OPENSSL_1_1_1}${LIBSSL_DIR}" "${BUILD_PYTHON_WITH_OPENSSL_1_1_1}" "${OPENSSL_LIBS_DIR}/1_1_1"
 # Copy ssl modules and libraries which are compiled for OpenSSL 3
-copy_and_patch_openssl_libs "${BUILD_OPENSSL_3}${LIBSSL_DIR}" "${BUILD_PYTHON_WITH_OPENSSL_3}" "${OPENSSL_LIBS_DIR}/3"
+copy_openssl_libs "${BUILD_OPENSSL_3}${LIBSSL_DIR}" "${BUILD_PYTHON_WITH_OPENSSL_3}" "${OPENSSL_LIBS_DIR}/3"
 
 
-# Patch Python executable and hardcode libpython.so, so runtime linker does not have to search for it.
-patchelf --replace-needed \
-  "libpython${PYTHON_SHORT_VERSION}.so.1.0" \
-  "${INSTALL_PREFIX}/lib/libpython${PYTHON_SHORT_VERSION}.so.1.0" \
-  "${PACKAGE_ROOT}${INSTALL_PREFIX}/bin/python${PYTHON_SHORT_VERSION}"
+RESULT_PYTHON_SSL_BINDINGS_DIR="${PACKAGE_ROOT}${INSTALL_PREFIX}/lib/python${PYTHON_SHORT_VERSION}/lib-dynload"
+rm "$(get_standard_c_binding_path "${RESULT_PYTHON_SSL_BINDINGS_DIR}" _ssl.cpython-*-*-*-*.so)"
+rm "$(get_standard_c_binding_path "${RESULT_PYTHON_SSL_BINDINGS_DIR}" _hashlib.cpython-*-*-*-*.so)"
 
 # Copy package scriptlets
 cp -a "${SOURCE_ROOT}/agent_build_refactored/managed_packages/scalyr_agent_python3/install-scriptlets" "${STEP_OUTPUT_PATH}/scriptlets"
 
-# Copy executable that allows configure the package.
-cp "${SOURCE_ROOT}/agent_build_refactored/managed_packages/scalyr_agent_python3/agent-python3-config" "${PACKAGE_ROOT}${INSTALL_PREFIX}/bin"
+BIN_DIR="${PACKAGE_ROOT}${INSTALL_PREFIX}/bin"
+# Copy executables that allows to configure the package.
+cp "${SOURCE_ROOT}/agent_build_refactored/managed_packages/scalyr_agent_python3/agent-python3-config" "${BIN_DIR}"
+cp -r "${SOURCE_ROOT}/agent_build_refactored/managed_packages/scalyr_agent_python3/internal" "${BIN_DIR}"
+
+# Rename main Python executable to be 'python3-original' and copy our wrapper script instead of it
+mv "${PACKAGE_ROOT}${INSTALL_PREFIX}/bin/python${PYTHON_SHORT_VERSION}" "${PACKAGE_ROOT}${INSTALL_PREFIX}/bin/python3-original"
+cp "${SOURCE_ROOT}/agent_build_refactored/managed_packages/scalyr_agent_python3/python3" "${BIN_DIR}/python${PYTHON_SHORT_VERSION}"
+
+# Remove other executables
+rm "${BIN_DIR}"/pip*
+rm "${BIN_DIR}"/idle*
+rm "${BIN_DIR}"/2to3*
+rm "${BIN_DIR}"/pydoc*
+
 
 # Copy package's configuration files.
 ETC_DIR="${PACKAGE_ROOT}${INSTALL_PREFIX}/etc"